# Performance improvements for AniDB_Anime cache initialisation