--- a/JMMServer/Collections/EnumerableExtensions.cs
+++ b/JMMServer/Collections/EnumerableExtensions.cs
@@ -1,0 +1,55 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace JMMServer.Collections
+{
+    public static class EnumerableExtensions
+    {
+        public static ILookup<TKey, TSource> ToLazyLookup<TKey, TSource>(this IEnumerable<TSource> source, Func<TSource, TKey> keySelector,
+            IEqualityComparer<TKey> comparer = null)
+        {
+            return LazyLookup<TKey, TSource>.Create(source, keySelector, comparer);
+        }
+
+        public static ILookup<TKey, TElement> ToLazyLookup<TSource, TKey, TElement>(this IEnumerable<TSource> source, Func<TSource, TKey> keySelector,
+            Func<TSource, TElement> elementSelector, IEqualityComparer<TKey> comparer = null)
+        {
+            return LazyLookup<TKey, TElement>.Create(source, keySelector, elementSelector, comparer);
+        }
+
+        public static IEnumerable<TSource[]> Batch<TSource>(this IEnumerable<TSource> source, int size)
+        {
+            TSource[] bucket = null;
+            var count = 0;
+
+            foreach (var item in source)
+            {
+                if (bucket == null)
+                {
+                    bucket = new TSource[size];
+                }
+
+                bucket[count++] = item;
+
+                if (count != size)
+                {
+                    continue;
+                }
+
+                yield return bucket;
+
+                bucket = null;
+                count = 0;
+            }
+
+            // Return the last bucket with remaining elements
+            if (bucket != null && count > 0)
+            {
+                Array.Resize(ref bucket, count);
+
+                yield return bucket;
+            }
+        }
+    }
+}