--- a/JMMServer/Databases/DatabaseFixes.cs
+++ b/JMMServer/Databases/DatabaseFixes.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using AniDBAPI;
 using JMMServer.Entities;
 using JMMServer.Repositories;
+using JMMServer.Repositories.NHibernate;
 using NLog;
 
 namespace JMMServer.Databases
@@ -212,6 +213,7 @@ namespace JMMServer.Databases
 
                 using (var session = JMMService.SessionFactory.OpenSession())
                 {
+                    ISessionWrapper sessionWrapper = session.Wrap();
                     List<CrossRef_AniDB_TvDB> xrefsTvDB = repCrossRefTvDB.GetAll();
                     foreach (CrossRef_AniDB_TvDB xrefTvDB in xrefsTvDB)
                     {
@@ -221,7 +223,7 @@ namespace JMMServer.Databases
                         xrefNew.TvDBID = xrefTvDB.TvDBID;
                         xrefNew.TvDBSeasonNumber = xrefTvDB.TvDBSeasonNumber;
 
-                        TvDB_Series ser = xrefTvDB.GetTvDBSeries(session);
+                        TvDB_Series ser = xrefTvDB.GetTvDBSeries(sessionWrapper);
                         if (ser != null)
                             xrefNew.TvDBTitle = ser.SeriesName;
 
@@ -265,7 +267,7 @@ namespace JMMServer.Databases
                         xrefNew.AniDBStartEpisodeType = (int) enEpisodeType.Special;
                         xrefNew.AniDBStartEpisodeNumber = 1;
 
-                        TvDB_Series ser = xrefTvDB.GetTvDBSeries(session);
+                        TvDB_Series ser = xrefTvDB.GetTvDBSeries(sessionWrapper);
                         if (ser != null)
                             xrefNew.TvDBTitle = ser.SeriesName;