--- a/JMMServer/Entities/IImageEntity.cs
+++ b/JMMServer/Entities/IImageEntity.cs
@@ -1,0 +1,6 @@
+﻿namespace JMMServer.Entities
+{
+    public interface IImageEntity
+    {
+    }
+}