--- a/JMMServer/Entities/MovieDB_Fanart.cs
+++ b/JMMServer/Entities/MovieDB_Fanart.cs
@@ -6,7 +6,7 @@ using NLog;
 
 namespace JMMServer.Entities
 {
-    public class MovieDB_Fanart
+    public class MovieDB_Fanart : IImageEntity
     {
         private static Logger logger = LogManager.GetCurrentClassLogger();