… JMMServer/Entities/AniDB_Anime.cs
… JMMServer/Entities/AnimeGroup.cs
… JMMServer/Entities/AnimeSeries.cs
… JMMServer/Entities/CrossRef_AniDB_TvDBV2.cs
… JMMServer/Entities/Trakt_ImageFanart.cs
… JMMServer/Entities/Trakt_ImagePoster.cs
… JMMServer/Entities/TvDB_ImageFanart.cs
… JMMServer/Entities/TvDB_ImagePoster.cs
… JMMServer/Entities/TvDB_ImageWideBanner.cs
… JMMServer/Entities/VideoLocal.cs
… JMMServer/Importer.cs
… JMMServer/JMMServer.csproj
… JMMServer/JMMServiceImplementation.cs
… JMMServer/JMMServiceImplementationMetro.cs
… JMMServer/PlexAndKodi/CommonImplementation.cs
… JMMServer/PlexAndKodi/Helper.cs
… JMMServer/Providers/MovieDB/MovieDBHelper.cs
… JMMServer/Providers/TvDB/TvDBHelper.cs
… JMMServer/Repositories/AdhocRepository.cs
… JMMServer/Repositories/AniDB_AnimeRepository.cs
… JMMServer/Repositories/AniDB_Anime_CharacterRepository.cs
… JMMServer/Repositories/AniDB_Anime_DefaultImageRepository.cs
… JMMServer/Repositories/AniDB_Anime_RelationRepository.cs
… JMMServer/Repositories/AniDB_Anime_TagRepository.cs
… JMMServer/Repositories/AniDB_Anime_TitleRepository.cs
… JMMServer/Repositories/AniDB_CharacterRepository.cs
… JMMServer/Repositories/AniDB_EpisodeRepository.cs
… JMMServer/Repositories/AniDB_FileRepository.cs
… JMMServer/Repositories/AniDB_RecommendationRepository.cs
… JMMServer/Repositories/AniDB_TagRepository.cs