After recently updating JMM Server, I encountered the cache initialization process which took several hours to complete on my media PC. So, I thought I'd see if I could improve the performance a bit.
Below is a quick summary of what I've done:

- AniDB_AnimeRepository.InitCache now uses a stateless session. Using the standard ISession has a large amount of overhead due to change tracking, etc. (the profiler I used indicated that approximately 80% of the time was spent in NHibernate's Commit method).
- The above change required the introduction of an ISessionWrapper to provide the common functionality from both ISession and IStatelessSession for performing SELECT queries in various parts of the system.
- AniDB_Anime processing was switch over to work in batches instead of one by one to reduce the number of round-trips to the database.

Below are the results of the change (Tested on an i7 2600k, 16GB RAM, SSD; with 3661 AniDB_Anime records in DB):

Database | Before | After
--- | --- | ---
**SQL Server (2014)** | 1h 02m 35s | 25s
**SQLite** | 2h 11m 30s | 27s
**MySQL (5.17.5)** | 1h 16m 12s | 2m 45s

Note that this PR only improves the cache initialization of the AniDB_Anime stage. I haven't attempted to tackle some of the other slow stages yet due to the amount of work involved. If I do make an attempt, I'll do it in a separate PR (assuming this one turns out ok).

Also, one bit that hasn't been fully tested is handling the default images in bulk. The database I'm working doesn't have any records for images (I think I must have turned that feature off), so it would be greatly appreciated if someone could give it a quick test (if not, I'll have to find some time to setup some test cases).