@da3dsoul 2016-09-08
In the future, try not to reformat the code, as it makes git throw a hissy calculating diff. I never had poor performance with this, but speedup is always welcome if it doesn't break anything. Do me a favor. I noticed you made a .wrap() to all of the sessions. Put that in a helper function so that we can call it as session and it forwards session.wrap(). @maxpiva what do think of this?