@da3dsoul 2016-09-08
Also, I am committing a fix to a previous issue that modifies many of those files, so check the code again after.