@pmcleish 2016-09-08
Sorry about the code reformat. It wasn't intentional, I think there was a file or two that Visual Studio complained about line endings and I clicked "Yes" out of habit to fix them.

I should have added more detail into the issue. It normally isn't a major problem when you start JMM Server, except for when the AniDB_Anime.ContractVersion is less than what it should be (e.g. like after a major JMM upgrade, etc). A friend and I both recently updated, and the cache initialization process took many hours for both of us (plus our collections are probably larger than average which wouldn't help).

I'm not quite sure I understand about the .Wrap() helper you want. I added .Wrap() as an extension method to both ISession and IStatelessSession to make it easy to use. Are you requesting something that converts ISessionWrapper to ISession?