@da3dsoul 2016-09-08
Nvm I see what you did. And I just now regenerated my AniDB anime contracts. Yes, it takes a long time. It'll be nice if this speeds up contract generation. Does this affect performance of retrieval at all?