@pmcleish 2016-09-08
General retrieval of contracts? no.
Every other operation is still using an ISession (But for the things I needed to change, it is simply wrapped by ISessionWrapper). So their performance remains the same.
I was going to attempt to improve performance of other areas as well, but it is a lot of work, so I just wanted to dip my toe in, so to speak, and get some other peoples thoughts.