@da3dsoul 2016-09-08
Ok I made my commit. Look over it and see if any changes need to be adapted to your code. I'll try and leave that section alone for a bit.