@ElementalCrisis 2016-09-08
Thanks for the PR! @maxpiva is one of our lead programmers so he'll have to review, in the meantime and hopefully going forward come join us on our Discord server! It's always a pleasure to meet our users. :)

https://discordapp.com/channels/96234011612958720/96234011612958720