@ElementalCrisis 2016-09-08
Hello @pmcleish, 

Looks like @da3dsoul commit's made the PR unable to be merged in, we talked it over in and would like to include this very nice addition to the last maintenance release for 3.6 which we're labeling 3.6.1.0 before we move on to the next version. So once the conflicts have been fixed, I'll be able to merge it in.